+++ CHANGELOG.md
@@ -31,7 +31,8 @@ All notable changes to the Wazuh app project will be documented in this file.
 
 ### Fixed
 
-- Fixed creation of log files [#3384](https://github.com/wazuh/wazuh-kibana-app/pull/3384) 
+- Fixed creation of log files [#3384](https://github.com/wazuh/wazuh-kibana-app/pull/3384)
+- Query config refactor [#3490](https://github.com/wazuh/wazuh-kibana-app/pull/3490)
 
 ## Wazuh v4.2.1 - Kibana 7.10.2 , 7.11.2 - Revision 4202